refactor: nodes dependent on siblings or children of siblings now use signals for loose coupling
almost all nodes use signals now for better composition. Only exception
are the states which inherit from playermovement state which defines its
owner as the player, since the state machine is for the player only and
is always directly under the player, therefore the logic is naturally
tightly coupled with the player.